SCREAM: moved LayoutType and get_layout_type to field_layout.*pp
It avoids circular dependencies when using get_layout_type
inside field.hpp.
